+++ Assets/02. Scripts/UI/ChildUI/UICraftItemSlot.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0f0059dd90c26bf448d4acbe5fe9028c
+guid: abecf588f53967b45ae739ae7e5ea97d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
