--- Assets/02. Scripts/UI/PopupUI/UIConfirmBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 141af15c8ab5e834b875b222a0e11457
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
