+++ Assets/02. Scripts/UI/PopupUI/UICooking.cs
@@ -1,11 +1,16 @@
 using System.Collections.Generic;
+using System.Linq;
+using UnityEngine;
 
 // 2024. 01. 16 Byun Jeongmin
 public class UICooking : UICraftBase
 {
+    [SerializeField] private GameObject _ingredientPrefab;
+
     public override void Awake()
     {
         base.Awake();
+        _itemPrefab = _ingredientPrefab;
     }
 
     public override void OnEnable()
@@ -13,9 +18,14 @@ public class UICooking : UICraftBase
         base.OnEnable();
     }
 
-    protected override UIConfirmBase GetConfirmPopup()
+    //protected override UIConfirmBase GetConfirmPopup()
+    //{
+    //    return Managers.UI.ShowPopupUI<UICookingConfirm>();
+    //}
+
+    protected override List<RecipeSO.Ingredient> GetRequiredDataList()
     {
-        return Managers.UI.ShowPopupUI<UICookingConfirm>();
+        return Managers.Data.cookingDataList.SelectMany(recipe => recipe.requiredItems).ToList();
     }
 
     protected override List<RecipeSO> GetDataList()
