+++ Assets/02. Scripts/UI/PopupUI/UIRecipe.cs
@@ -1,11 +1,16 @@
 using System.Collections.Generic;
+using System.Linq;
+using UnityEngine;
 
 // 2024. 01. 11 Byun Jeongmin
 public class UIRecipe : UICraftBase
 {
+    [SerializeField] private GameObject _materialPrefab;
+
     public override void Awake()
     {
         base.Awake();
+        _itemPrefab = _materialPrefab;
     }
 
     public override void OnEnable()
@@ -17,16 +22,21 @@ public class UIRecipe : UICraftBase
     // 고급 레시피 UI의 활성화 여부를 설정하는 메서드
     public void SetAdvancedRecipeUIActive(bool active)
     {
-        foreach (var slot in _uiRecipeSlots)
+        foreach (var slot in _uiCraftSlots)
         {
             bool isAdvancedRecipe = Managers.Data.recipeDataList[slot.Index].IsAdvancedRecipe;
             slot.gameObject.SetActive(active || (!isAdvancedRecipe && !active));
         }
     }
 
-    protected override UIConfirmBase GetConfirmPopup()
+    //protected override UIConfirmBase GetConfirmPopup()
+    //{
+    //    return Managers.UI.ShowPopupUI<UIRecipeConfirm>();
+    //}
+
+    protected override List<RecipeSO.Ingredient> GetRequiredDataList()
     {
-        return Managers.UI.ShowPopupUI<UIRecipeConfirm>();
+        return (Managers.Data.recipeDataList.SelectMany(recipe => recipe.requiredItems).ToList());
     }
 
     protected override List<RecipeSO> GetDataList()
