--- Assets/02. Scripts/UI/PopupUI/UIRecipeConfirm.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 65019b4ac0d991e4792843671b1d2295
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
