+++ Assets/04. Resources/UI/Child/UICraftItemSlot.prefab.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+<<<<<<<< HEAD:Assets/04. Resources/UI/UIQuest.prefab.meta
+guid: 8f3d58d11d2783641b4b41f4e385e3f5
+PrefabImporter:
+========
+guid: 2e2e7987ac64a0c448b786d311ab0a8b
+NativeFormatImporter:
+>>>>>>>> Develop1.0:Assets/04. Resources/Prefabs/Icons/Wood.mat.meta
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
