+++ Assets/04. Resources/UI/Child/UITutorial.prefab
@@ -37,7 +37,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 0.5}
   m_AnchorMax: {x: 0, y: 0.5}
   m_AnchoredPosition: {x: -190, y: 5}
-  m_SizeDelta: {x: 500, y: 290}
+  m_SizeDelta: {x: 380, y: 240}
   m_Pivot: {x: 0, y: 0.5}
 --- !u!222 &8974081605171387521
 CanvasRenderer:
@@ -593,7 +593,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
-  m_SizeDelta: {x: 450, y: 220}
+  m_SizeDelta: {x: 330, y: 150}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &2492416730151085012
 CanvasRenderer:
@@ -870,7 +870,7 @@ RectTransform:
   m_AnchorMin: {x: 0, y: 0.5}
   m_AnchorMax: {x: 0, y: 0.5}
   m_AnchoredPosition: {x: -165, y: -17}
-  m_SizeDelta: {x: 450, y: 220}
+  m_SizeDelta: {x: 330, y: 150}
   m_Pivot: {x: 0, y: 0.5}
 --- !u!222 &7254342503998741526
 CanvasRenderer:
