--- Assets/04. Resources/UI/Popup/UICookingConfirm.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8ba5486169438f847a63142d0981472c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
