--- Assets/04. Resources/UI/Popup/UIRecipeConfirm.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0822b9175e47fac4ab1e88cd4574b36b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
